Added a bigram search, improved the PlantDetailsPage, and also the ability to access it through the plant search in the main menu.

+++ PlantApp/ViewModels/PlantDetailsViewModel.cs
@@ -3,54 +3,53 @@ using CommunityToolkit.Mvvm.Input;
 using Microsoft.EntityFrameworkCore;
 using PlantApp.Data;
 using PlantApp.Services;
+using System.Globalization;
 
-public partial class PlantDetailsViewModel : ObservableObject
+public partial class PlantDetailsViewModel : ObservableObject, IInitialize<Plant>
 {
+    [ObservableProperty] private Plant plant;
+    private int _plantId;
     private readonly IDbContextFactory<AppDbContext> _factory;
     private readonly AuthService _authService;
 
-    public Plant Plant { get; }
-    private readonly int _plantId;
-
-    public PlantDetailsViewModel(
-        IDbContextFactory<AppDbContext> factory,
-        Plant plant,
-        AuthService authService)
+    public PlantDetailsViewModel(IDbContextFactory<AppDbContext> factory, AuthService authService)
     {
         _factory = factory;
         _authService = authService;
+    }
 
+    public void Initialize(Plant plant)
+    {
         Plant = plant;
-        _plantId = plant.Id;   // сохраняем ID
+        _plantId = plant.Id;
     }
 
     [RelayCommand]
     public async Task ToggleFavorite()
     {
         using var db = _factory.CreateDbContext();
-
-        int currentUserId = _authService.GetUserId();
+        int userId = _authService.GetUserId();
 
         var favorite = await db.FavoritePlants
-            .FirstOrDefaultAsync(f =>
-                f.PlantId == _plantId &&
-                f.UserId == currentUserId);
+            .FirstOrDefaultAsync(f => f.PlantId == _plantId && f.UserId == userId);
 
-        if (favorite != null)
-        {
-            db.FavoritePlants.Remove(favorite);
-        }
-        else
-        {
-            var newFavorite = new FavoritePlant
-            {
-                PlantId = _plantId,
-                UserId = currentUserId
-            };
-
-            db.FavoritePlants.Add(newFavorite);
-        }
+        if (favorite != null) db.FavoritePlants.Remove(favorite);
+        else db.FavoritePlants.Add(new FavoritePlant { PlantId = _plantId, UserId = userId });
 
         await db.SaveChangesAsync();
     }
+
+    public class BoolToStarConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return (value is bool b && b) ? "⭐" : "☆";
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+
 }
